chore: refactor: move gittest and registrytest to testutil
